# apps/libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@boxlite.com
    name: BoxLite Platforms Inc.
    url: https://www.boxlite.io
  description: BoxLite AI platform API Docs
  title: BoxLite
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /admin/infrastructure-logs/access:
    get:
      operationId: adminCheckInfrastructureLogsAccess
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureLogsAccessDto"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Check infrastructure log access
      tags:
      - admin
  /admin/infrastructure-logs:
    get:
      operationId: adminSearchInfrastructureLogs
      parameters:
      - explode: true
        in: query
        name: source
        required: false
        schema:
          default: runner
          enum:
          - runner
          - collector
          type: string
        style: form
      - explode: true
        in: query
        name: from
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - explode: true
        in: query
        name: to
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - description: Case-sensitive literal phrase to find in a log message
        explode: true
        in: query
        name: search
        required: false
        schema:
          maxLength: 256
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          type: number
        style: form
      - description: Opaque CloudWatch pagination cursor
        explode: true
        in: query
        name: nextToken
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureLogs"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Search infrastructure fallback logs
      tags:
      - admin
  /admin/infrastructure-logs/platform:
    get:
      operationId: adminSearchPlatformLogs
      parameters:
      - description: Start of time range (ISO 8601)
        explode: true
        in: query
        name: from
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - description: End of time range (ISO 8601)
        explode: true
        in: query
        name: to
        required: true
        schema:
          format: date-time
          type: string
        style: form
      - description: Page number (1-indexed)
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          maximum: 10000
          minimum: 1
          type: number
        style: form
      - description: Number of items per page
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          type: number
        style: form
      - description: "Filter by severity levels (DEBUG, INFO, WARN, ERROR)"
        explode: true
        in: query
        name: severities
        required: false
        schema:
          items:
            type: string
          type: array
        style: form
      - description: Case-insensitive text search in the log body
        explode: true
        in: query
        name: search
        required: false
        schema:
          maxLength: 256
          type: string
        style: form
      - explode: true
        in: query
        name: source
        required: false
        schema:
          default: api
          enum:
          - api
          - worker
          - runner
          - box
          type: string
        style: form
      - description: Exact Box ID. Required when source is box.
        explode: true
        in: query
        name: boxId
        required: false
        schema:
          maxLength: 128
          type: string
        style: form
      - description: Exact OpenTelemetry trace ID
        explode: true
        in: query
        name: traceId
        required: false
        schema:
          maxLength: 32
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLogs"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Search allowlisted platform OTLP logs
      tags:
      - admin
...
components:
  schemas:
...
    InfrastructureLogsAccessDto:
      example:
        canRead: true
      properties:
        canRead:
          type: boolean
      required:
      - canRead
      type: object
    LogEntry:
      example:
        timestamp: timestamp
        body: body
        severityText: severityText
        severityNumber: 0.8008281904610115
        serviceName: serviceName
        resourceAttributes:
          key: resourceAttributes
        logAttributes:
          key: logAttributes
        traceId: traceId
        spanId: spanId
      properties:
        timestamp:
          description: Timestamp of the log entry
          type: string
        body:
          description: Log message body
          type: string
        severityText:
          description: "Severity level text (e.g., INFO, WARN, ERROR)"
          type: string
        severityNumber:
          description: Severity level number
          type: number
        serviceName:
          description: Service name that generated the log
          type: string
        resourceAttributes:
          additionalProperties:
            type: string
          description: Resource attributes from OTEL
          type: object
        logAttributes:
          additionalProperties:
            type: string
          description: Log-specific attributes
          type: object
        traceId:
          description: Associated trace ID if available
          type: string
        spanId:
          description: Associated span ID if available
          type: string
      required:
      - body
      - logAttributes
      - resourceAttributes
      - serviceName
      - severityText
      - timestamp
      type: object
    InfrastructureLogs:
      example:
        items:
        - timestamp: timestamp
          body: body
          severityText: severityText
          severityNumber: 0.8008281904610115
          serviceName: serviceName
          resourceAttributes:
            key: resourceAttributes
          logAttributes:
            key: logAttributes
          traceId: traceId
          spanId: spanId
        - timestamp: timestamp
          body: body
          severityText: severityText
          severityNumber: 0.8008281904610115
          serviceName: serviceName
          resourceAttributes:
            key: resourceAttributes
          logAttributes:
            key: logAttributes
          traceId: traceId
          spanId: spanId
        nextToken: nextToken
      properties:
        items:
          items:
            $ref: "#/components/schemas/LogEntry"
          type: array
        nextToken:
          description: Opaque cursor for the next CloudWatch page
          type: string
      required:
      - items
      type: object
    PaginatedLogs:
      example:
        items:
        - timestamp: timestamp
          body: body
          severityText: severityText
          severityNumber: 0.8008281904610115
          serviceName: serviceName
          resourceAttributes:
            key: resourceAttributes
          logAttributes:
            key: logAttributes
          traceId: traceId
          spanId: spanId
        - timestamp: timestamp
          body: body
          severityText: severityText
          severityNumber: 0.8008281904610115
          serviceName: serviceName
          resourceAttributes:
            key: resourceAttributes
          logAttributes:
            key: logAttributes
          traceId: traceId
          spanId: spanId
        total: 0.8008281904610115
        page: 6.027456183070403
        totalPages: 1.4658129805029452
      properties:
        items:
          description: List of log entries
          items:
            $ref: "#/components/schemas/LogEntry"
          type: array
        total:
          description: Total number of log entries matching the query
          type: number
        page:
          description: Current page number
          type: number
        totalPages:
          description: Total number of pages
          type: number
      required:
      - items
      - page
      - total
      - totalPages
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect