# apps/libs/api-client-go/model_infrastructure_logs.go
/*
BoxLite

BoxLite AI platform API Docs

API version: 1.0
Contact: support@boxlite.com
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package apiclient

import (
	"encoding/json"
	"fmt"
)

// checks if the InfrastructureLogs type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &InfrastructureLogs{}

// InfrastructureLogs struct for InfrastructureLogs
type InfrastructureLogs struct {
	Items []LogEntry `json:"items"`
	// Opaque cursor for the next CloudWatch page
	NextToken *string `json:"nextToken,omitempty"`
	AdditionalProperties map[string]interface{}
}

type _InfrastructureLogs InfrastructureLogs

// NewInfrastructureLogs instantiates a new InfrastructureLogs object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewInfrastructureLogs(items []LogEntry) *InfrastructureLogs {
	this := InfrastructureLogs{}
	this.Items = items
	return &this
}

// NewInfrastructureLogsWithDefaults instantiates a new InfrastructureLogs object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewInfrastructureLogsWithDefaults() *InfrastructureLogs {
	this := InfrastructureLogs{}
	return &this
}

// GetItems returns the Items field value
func (o *InfrastructureLogs) GetItems() []LogEntry {
	if o == nil {
		var ret []LogEntry
		return ret
	}

	return o.Items
}

// GetItemsOk returns a tuple with the Items field value
// and a boolean to check if the value has been set.
func (o *InfrastructureLogs) GetItemsOk() ([]LogEntry, bool) {
	if o == nil {
		return nil, false
	}
	return o.Items, true
}

// SetItems sets field value
func (o *InfrastructureLogs) SetItems(v []LogEntry) {
	o.Items = v
}

// GetNextToken returns the NextToken field value if set, zero value otherwise.
func (o *InfrastructureLogs) GetNextToken() string {
	if o == nil || IsNil(o.NextToken) {
		var ret string
		return ret
	}
	return *o.NextToken
}

// GetNextTokenOk returns a tuple with the NextToken field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *InfrastructureLogs) GetNextTokenOk() (*string, bool) {
	if o == nil || IsNil(o.NextToken) {
		return nil, false
	}
	return o.NextToken, true
}

// HasNextToken returns a boolean if a field has been set.
func (o *InfrastructureLogs) HasNextToken() bool {
	if o != nil && !IsNil(o.NextToken) {
		return true
	}

	return false
}

// SetNextToken gets a reference to the given string and assigns it to the NextToken field.
func (o *InfrastructureLogs) SetNextToken(v string) {
	o.NextToken = &v
}

func (o InfrastructureLogs) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o InfrastructureLogs) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["items"] = o.Items
	if !IsNil(o.NextToken) {
		toSerialize["nextToken"] = o.NextToken
	}

	for key, value := range o.AdditionalProperties {
		toSerialize[key] = value
	}

	return toSerialize, nil
}

func (o *InfrastructureLogs) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"items",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err;
	}

	for _, requiredProperty := range(requiredProperties) {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varInfrastructureLogs := _InfrastructureLogs{}

	err = json.Unmarshal(data, &varInfrastructureLogs)

	if err != nil {
		return err
	}

	*o = InfrastructureLogs(varInfrastructureLogs)

	additionalProperties := make(map[string]interface{})

	if err = json.Unmarshal(data, &additionalProperties); err == nil {
		delete(additionalProperties, "items")
		delete(additionalProperties, "nextToken")
		o.AdditionalProperties = additionalProperties
	}

	return err
}

type NullableInfrastructureLogs struct {
	value *InfrastructureLogs
	isSet bool
}

func (v NullableInfrastructureLogs) Get() *InfrastructureLogs {
	return v.value
}

func (v *NullableInfrastructureLogs) Set(val *InfrastructureLogs) {
	v.value = val
	v.isSet = true
}

func (v NullableInfrastructureLogs) IsSet() bool {
	return v.isSet
}

func (v *NullableInfrastructureLogs) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableInfrastructureLogs(val *InfrastructureLogs) *NullableInfrastructureLogs {
	return &NullableInfrastructureLogs{value: val, isSet: true}
}

func (v NullableInfrastructureLogs) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableInfrastructureLogs) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}