# apps/libs/api-client-go/model_infrastructure_logs_access_dto.go
/*
BoxLite

BoxLite AI platform API Docs

API version: 1.0
Contact: support@boxlite.com
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package apiclient

import (
	"encoding/json"
	"fmt"
)

// checks if the InfrastructureLogsAccessDto type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &InfrastructureLogsAccessDto{}

// InfrastructureLogsAccessDto struct for InfrastructureLogsAccessDto
type InfrastructureLogsAccessDto struct {
	CanRead bool `json:"canRead"`
	AdditionalProperties map[string]interface{}
}

type _InfrastructureLogsAccessDto InfrastructureLogsAccessDto

// NewInfrastructureLogsAccessDto instantiates a new InfrastructureLogsAccessDto object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewInfrastructureLogsAccessDto(canRead bool) *InfrastructureLogsAccessDto {
	this := InfrastructureLogsAccessDto{}
	this.CanRead = canRead
	return &this
}

// NewInfrastructureLogsAccessDtoWithDefaults instantiates a new InfrastructureLogsAccessDto object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewInfrastructureLogsAccessDtoWithDefaults() *InfrastructureLogsAccessDto {
	this := InfrastructureLogsAccessDto{}
	return &this
}

// GetCanRead returns the CanRead field value
func (o *InfrastructureLogsAccessDto) GetCanRead() bool {
	if o == nil {
		var ret bool
		return ret
	}

	return o.CanRead
}

// GetCanReadOk returns a tuple with the CanRead field value
// and a boolean to check if the value has been set.
func (o *InfrastructureLogsAccessDto) GetCanReadOk() (*bool, bool) {
	if o == nil {
		return nil, false
	}
	return &o.CanRead, true
}

// SetCanRead sets field value
func (o *InfrastructureLogsAccessDto) SetCanRead(v bool) {
	o.CanRead = v
}

func (o InfrastructureLogsAccessDto) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o InfrastructureLogsAccessDto) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["canRead"] = o.CanRead

	for key, value := range o.AdditionalProperties {
		toSerialize[key] = value
	}

	return toSerialize, nil
}

func (o *InfrastructureLogsAccessDto) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"canRead",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err;
	}

	for _, requiredProperty := range(requiredProperties) {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varInfrastructureLogsAccessDto := _InfrastructureLogsAccessDto{}

	err = json.Unmarshal(data, &varInfrastructureLogsAccessDto)

	if err != nil {
		return err
	}

	*o = InfrastructureLogsAccessDto(varInfrastructureLogsAccessDto)

	additionalProperties := make(map[string]interface{})

	if err = json.Unmarshal(data, &additionalProperties); err == nil {
		delete(additionalProperties, "canRead")
		o.AdditionalProperties = additionalProperties
	}

	return err
}

type NullableInfrastructureLogsAccessDto struct {
	value *InfrastructureLogsAccessDto
	isSet bool
}

func (v NullableInfrastructureLogsAccessDto) Get() *InfrastructureLogsAccessDto {
	return v.value
}

func (v *NullableInfrastructureLogsAccessDto) Set(val *InfrastructureLogsAccessDto) {
	v.value = val
	v.isSet = true
}

func (v NullableInfrastructureLogsAccessDto) IsSet() bool {
	return v.isSet
}

func (v *NullableInfrastructureLogsAccessDto) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableInfrastructureLogsAccessDto(val *InfrastructureLogsAccessDto) *NullableInfrastructureLogsAccessDto {
	return &NullableInfrastructureLogsAccessDto{value: val, isSet: true}
}

func (v NullableInfrastructureLogsAccessDto) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableInfrastructureLogsAccessDto) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}