# apps/libs/api-client/src/models/infrastructure-logs-access-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BoxLite
 * BoxLite AI platform API Docs
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@boxlite.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface InfrastructureLogsAccessDto {
    'canRead': boolean;
}