# apps/libs/api-client/src/models/infrastructure-logs.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BoxLite
 * BoxLite AI platform API Docs
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@boxlite.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { LogEntry } from './log-entry';

export interface InfrastructureLogs {
    'items': Array<LogEntry>;
    /**
     * Opaque cursor for the next CloudWatch page
     */
    'nextToken'?: string;
}